# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payout-catalog: {"get":{"operationId":"getPayoutCatalog","tags":["payout-catalog"],"security":[{"sessionBearer":[]}],"responses":{"200":{"description":"Success; observed pricing is an estimate requiring a fresh provider quote","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayoutCatalog"}}}},"400":{"description":"Invalid input"},"401":{"description":"Authentication required"},"403":{"description":"Enabled owner and production MFA required for admin operations"},"404":{"description":"Catalog entry unavailable"},"409":{"description":"Conflicting observation or outdated revision"},"503":{"description":"Catalog unavailable or account scope not configured"}},"summary":"Selectable provider → country → method hierarchy; observed send amounts and fees remain estimates"}}
  /payout-catalog/estimate: {"post":{"operationId":"estimatePayoutMethod","tags":["payout-catalog"],"security":[{"sessionBearer":[]}],"responses":{"200":{"description":"Success; observed pricing is an estimate requiring a fresh provider quote","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayoutEstimate"}}}},"400":{"description":"Invalid input"},"401":{"description":"Authentication required"},"403":{"description":"Enabled owner and production MFA required for admin operations"},"404":{"description":"Catalog entry unavailable"},"409":{"description":"Conflicting observation or outdated revision"},"503":{"description":"Catalog unavailable or account scope not configured"}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayoutEstimateRequest"}}}},"summary":"Estimate an exact observed amount within USD 15 gross; missing matching quotes return a null breakdown"}}
  /admin-data/payout-catalog: {"get":{"operationId":"getAdminPayoutCatalog","tags":["payout-catalog"],"security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success; observed pricing is an estimate requiring a fresh provider quote","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayoutCatalog"}}}},"400":{"description":"Invalid input"},"401":{"description":"Authentication required"},"403":{"description":"Enabled owner and production MFA required for admin operations"},"404":{"description":"Catalog entry unavailable"},"409":{"description":"Conflicting observation or outdated revision"},"503":{"description":"Catalog unavailable or account scope not configured"}},"summary":"Read complete account-scoped catalog including unavailable and errored routes"}}
  /admin-data/payout-catalog/import: {"post":{"operationId":"importPayoutResearch","tags":["payout-catalog"],"security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success; observed pricing is an estimate requiring a fresh provider quote","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayoutResearchImportResult"}}}},"400":{"description":"Invalid input"},"401":{"description":"Authentication required"},"403":{"description":"Enabled owner and production MFA required for admin operations"},"404":{"description":"Catalog entry unavailable"},"409":{"description":"Conflicting observation or outdated revision"},"503":{"description":"Catalog unavailable or account scope not configured"}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayoutResearchImportRequest"}}}},"summary":"Validate or atomically import signed-in Remitly Business research, idempotently"}}
  /admin-data/payout-catalog/{kind}/{id}: {"patch":{"operationId":"updatePayoutCatalogEntry","tags":["payout-catalog"],"security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success; observed pricing is an estimate requiring a fresh provider quote","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayoutCatalogEntry"}}}},"400":{"description":"Invalid input"},"401":{"description":"Authentication required"},"403":{"description":"Enabled owner and production MFA required for admin operations"},"404":{"description":"Catalog entry unavailable"},"409":{"description":"Conflicting observation or outdated revision"},"503":{"description":"Catalog unavailable or account scope not configured"}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayoutCatalogUpdate"}}}},"summary":"Update display name or enablement with revision and audit history","parameters":[{"name":"kind","in":"path","required":true,"schema":{"type":"string","enum":["providers","countries","methods"]}},{"name":"id","in":"path","required":true,"schema":{"type":"string","maxLength":250}}]}}
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
    adminBearer:
      type: http
      scheme: bearer
  schemas:
...
    PayoutFeeObservation: {"type":"object","required":["id","observedAt","sendAmountCents","feeCents","feeCurrency","fundingMethod","senderCountry","deliveryEstimate","taxStatus","liveRequoteRequired"],"properties":{"id":{"type":"string"},"observedAt":{"type":"string","format":"date-time"},"sendAmountCents":{"type":"integer","minimum":0},"feeCents":{"type":"integer","minimum":0},"feeCurrency":{"type":"string","enum":["USD"]},"fundingMethod":{"type":"string"},"senderCountry":{"type":"string"},"deliveryEstimate":{"type":"string","nullable":true},"taxStatus":{"type":"string","enum":["unresolved","not_observed"]},"liveRequoteRequired":{"type":"boolean"},"inspectionStatus":{"type":"string"},"linkMethodVerified":{"type":"boolean"},"sourceUrls":{"type":"array","items":{"type":"string"}},"discountNote":{"type":"string","nullable":true},"discountVerified":{"type":"boolean"}}}
    PayoutCountryObservation: {"type":"object","required":["id","observedAt","availability","inspectionStatus","notes","sourceUrls","discountNote","discountVerified","scope","receiveCurrency","senderCountry","fundingMethod","liveRequoteRequired"],"properties":{"id":{"type":"string"},"observedAt":{"type":"string","format":"date-time"},"availability":{"type":"string","enum":["available","unavailable","quote_error","unverified"]},"inspectionStatus":{"type":"string"},"notes":{"type":"string"},"sourceUrls":{"type":"array","items":{"type":"string"}},"discountNote":{"type":"string","nullable":true},"discountVerified":{"type":"boolean"},"scope":{"type":"string"},"receiveCurrency":{"type":"string","nullable":true},"senderCountry":{"type":"string"},"fundingMethod":{"type":"string"},"liveRequoteRequired":{"type":"boolean"}}}
    PayoutDeliveryMethod: {"type":"object","required":["id","name","enabled","revision","code","receiveCurrency","availability","lastVerifiedAt","observations"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"enabled":{"type":"boolean"},"revision":{"type":"integer","minimum":1},"code":{"type":"string"},"receiveCurrency":{"type":"string"},"availability":{"type":"string","enum":["available","unavailable","quote_error","unverified"]},"lastVerifiedAt":{"type":"string","format":"date-time"},"observations":{"type":"array","items":{"$ref":"#/components/schemas/PayoutFeeObservation"}}}}
    PayoutCountry: {"type":"object","required":["id","name","enabled","revision","countryCode","availability","lastVerifiedAt","methods"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"enabled":{"type":"boolean"},"revision":{"type":"integer","minimum":1},"countryCode":{"type":"string"},"availability":{"type":"string","enum":["available","unavailable","quote_error","unverified"]},"lastVerifiedAt":{"type":"string","format":"date-time"},"methods":{"type":"array","items":{"$ref":"#/components/schemas/PayoutDeliveryMethod"}},"observations":{"type":"array","items":{"$ref":"#/components/schemas/PayoutCountryObservation"}}}}
    PayoutProvider: {"type":"object","required":["id","name","enabled","revision","countries"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"enabled":{"type":"boolean"},"revision":{"type":"integer","minimum":1},"countries":{"type":"array","items":{"$ref":"#/components/schemas/PayoutCountry"}},"accountKey":{"type":"string"}}}
    PayoutCatalog: {"type":"object","required":["providers","maxWithdrawalCents","liveRequoteRequired","asOf","environment"],"properties":{"providers":{"type":"array","items":{"$ref":"#/components/schemas/PayoutProvider"}},"maxWithdrawalCents":{"type":"integer","enum":[1500]},"liveRequoteRequired":{"type":"boolean"},"asOf":{"type":"string","format":"date-time"},"environment":{"type":"string","enum":["development","production"]}}}
    PayoutEstimateRequest: {"type":"object","required":["methodId","withdrawalCents"],"properties":{"methodId":{"type":"string"},"withdrawalCents":{"type":"integer","minimum":1,"maximum":1500},"fundingMethod":{"type":"string","default":"debit_card"}},"additionalProperties":false}
    PayoutEstimateBreakdown: {"type":"object","required":["feeCents","sendAmountCents","totalEarningsDeductedCents","feeCurrency","taxCents","recipientAmount","observedAt"],"properties":{"feeCents":{"type":"integer","minimum":0},"sendAmountCents":{"type":"integer","minimum":0},"totalEarningsDeductedCents":{"type":"integer","minimum":0},"feeCurrency":{"type":"string","enum":["USD"]},"taxCents":{"type":"integer","nullable":true},"recipientAmount":{"type":"string","nullable":true},"observedAt":{"type":"string","format":"date-time"}}}
    PayoutEstimate: {"type":"object","required":["methodId","withdrawalCents","fundingMethod","receiveCurrency","quoteRequired","liveRequoteRequired","breakdown","message"],"properties":{"methodId":{"type":"string"},"withdrawalCents":{"type":"integer","minimum":0},"fundingMethod":{"type":"string"},"receiveCurrency":{"type":"string"},"quoteRequired":{"type":"boolean"},"liveRequoteRequired":{"type":"boolean"},"breakdown":{"$ref":"#/components/schemas/PayoutEstimateBreakdown","nullable":true},"message":{"type":"string"}}}
    PayoutCatalogUpdate: {"type":"object","required":["revision"],"properties":{"revision":{"type":"integer","minimum":1},"name":{"type":"string","minLength":1,"maxLength":160},"enabled":{"type":"boolean"}},"additionalProperties":false}
    PayoutCatalogEntry: {"type":"object","required":["id","name","enabled","revision"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"enabled":{"type":"boolean"},"revision":{"type":"integer","minimum":1}}}
    PayoutResearch: {"type":"object","required":["observed_date","source","source_urls","sender_country","funding_method","comparison_send_amounts_usd","fee_currency","production_fee_schedule","live_requote_required","scope","countries"],"properties":{"observed_date":{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},"source":{"type":"string","enum":["Signed-in Remitly Business website UI"]},"source_urls":{"type":"array","items":{"type":"string"}},"sender_country":{"type":"string","enum":["US"]},"funding_method":{"type":"string","enum":["debit_card","credit_card","bank_account"]},"comparison_send_amounts_usd":{"type":"array","items":{"type":"number","minimum":0.01,"maximum":100000}},"fee_currency":{"type":"string","enum":["USD"]},"production_fee_schedule":{"type":"boolean","enum":[false]},"live_requote_required":{"type":"boolean","enum":[true]},"scope":{"type":"string"},"discount_note":{"type":"string","nullable":true},"countries":{"type":"array","items":{"$ref":"#/components/schemas/PayoutResearchCountry"}},"observed_at":{"type":"string","format":"date-time"}}}
    PayoutResearchCountry: {"type":"object","required":["country_code","country","receive_currency","inspection_status","notes","methods"],"properties":{"country_code":{"type":"string","pattern":"^[A-Z]{2}$"},"country":{"type":"string"},"receive_currency":{"type":"string","nullable":true},"inspection_status":{"type":"string","enum":["link_options","manual_only","not_in_destination_picker","quote_error","link_available_no_method_modal"]},"notes":{"type":"string"},"methods":{"type":"array","items":{"type":"object","required":["label","delivery_estimate"],"properties":{"label":{"type":"string"},"delivery_estimate":{"type":"string","nullable":true}},"additionalProperties":{"oneOf":[{"type":"integer","minimum":0},{"type":"string","nullable":true}]}}}}}
    PayoutResearchImportRequest: {"type":"object","required":["research"],"properties":{"research":{"$ref":"#/components/schemas/PayoutResearch"},"dryRun":{"type":"boolean","default":true}},"additionalProperties":false}
    PayoutResearchImportResult: {"type":"object","required":["dryRun","providerId","countries","methods","observations"],"properties":{"dryRun":{"type":"boolean"},"providerId":{"type":"string"},"countries":{"type":"integer"},"methods":{"type":"integer"},"observations":{"type":"integer"},"insertedObservations":{"type":"integer"}}}
...