# lib/api-zod/src/generated/types/payoutEstimate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PayoutEstimateBreakdown } from './payoutEstimateBreakdown';

export interface PayoutEstimate {
  methodId: string;
  /** @minimum 0 */
  withdrawalCents: number;
  fundingMethod: string;
  receiveCurrency: string;
  quoteRequired: boolean;
  liveRequoteRequired: boolean;
  breakdown: PayoutEstimateBreakdown | null;
  message: string;
}